+++ packages/react-impression-examples/src/list/container.js
@@ -1,8 +1,7 @@
 import React, { Component } from 'react'
 import Option from './components/Option'
-import Edit from './components/Edit'
-import { Table, Confirm, Split, Card, Button } from 'react-impression'
 import { data, cityList } from './config'
+import TableContainer from './containers/TableContainer'
 import './container.module.scss'
 
 class List extends Component {
@@ -13,8 +12,6 @@ class List extends Component {
       activePage: 1,
     },
     option: {},
-    showEdit: false,
-    showDelete: false,
   }
 
   setOption = (path, value) => {
@@ -26,62 +23,10 @@ class List extends Component {
     })
   }
 
-  toggleShowEdit = () => {
-    this.setState({
-      showEdit: !this.state.showEdit,
-    })
-  }
-
-  toggleShowDelete = () => {
-    this.setState({
-      showDelete: !this.state.showDelete,
-    })
-  }
-
   searchList = () => {}
 
   render() {
-    const { list, showEdit, showDelete } = this.state
-
-    const columns = [
-      {
-        prop: 'id',
-        Header: ' 任务ID',
-      },
-      { prop: 'cityName', Header: '城市' },
-      { prop: 'taskName', Header: '任务名称', width: 120 },
-      { prop: 'source', Header: '数据来源' },
-      { prop: 'type', Header: '任务类型' },
-      { prop: 'taskObject', Header: '任务对象' },
-      { prop: 'result', Header: '任务结果' },
-      { prop: 'date', Header: '任务期间' },
-      { prop: 'num', Header: '任务对象数' },
-      {
-        prop: 'id',
-        Header: '操作',
-        width: 150,
-        Cell: value => {
-          return (
-            <div className='text-center'>
-              <a href='javascript:;' onClick={this.toggleShowEdit}>
-                编辑
-              </a>
-              <Split />
-              <a href='javascript:;' onClick={this.toggleShowDelete}>
-                删除
-              </a>
-            </div>
-          )
-        },
-      },
-    ]
-
-    const pagination = {
-      onSelect: this.searchList,
-      activePage: list.activePage,
-      totalPage: list.totalPage,
-      scope: 4,
-    }
+    const { list } = this.state
 
     return (
       <div styleName='container'>
@@ -90,25 +35,7 @@ class List extends Component {
           set={this.setOption}
           search={this.searchList}
         />
-        <Card block>
-          <Card outline='none'>
-            <Button theme='default'>启用</Button>
-            <Button theme='default' className='offset-l-lg'>
-              停用
-            </Button>
-          </Card>
-          <Table data={list.result} columns={columns} pagination={pagination} />
-        </Card>
-        <Edit show={showEdit} toggle={this.toggleShowEdit} />
-        {showDelete && (
-          <Confirm
-            type='danger'
-            onOkClick={this.toggleShowDelete}
-            onCancelClick={this.toggleShowDelete}
-          >
-            您确定删除该记录？
-          </Confirm>
-        )}
+        <TableContainer list={list} search={this.searchList} />
       </div>
     )
   }
